Moved calls into their own section
When including the same call in two places with toctrees, then the
second place links to the document in the first place.

Example:
`session-post` is included in `CPOs -> CDRs` and `EMPs -> CDRs`.
Now when a user clicks on the `session-post` link on `EMPs`, it takes
the user to the documentation at `CDRs`.

This is not intuitive for the user, thus the calls are split into their
own section. The explanations for CPOs and EMPs link to the calls
via `:ref:`. I think it is also nice for the user to have a list off
all possible calls in some place.